backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
python-dotenv
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers